Move request and response DTOs to dto package
