+++ src/main/java/com/ite/itea/presentation/CheckoutController.java
@@ -3,9 +3,9 @@ package com.ite.itea.presentation;
 import com.ite.itea.application.dto.*;
 import com.ite.itea.application.usecase.CalculateTotalPriceUseCase;
 import com.ite.itea.domain.retail.ProductName;
-import com.ite.itea.presentation.request.ItemRequest;
-import com.ite.itea.presentation.request.OrderRequest;
-import com.ite.itea.presentation.response.ReceiptResponse;
+import com.ite.itea.application.dto.ItemRequest;
+import com.ite.itea.application.dto.OrderRequest;
+import com.ite.itea.application.dto.ReceiptResponse;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
